// src/generated/models/PeriodicTaskMutationRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Codebook Api
 * A simple api
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { CeleryPeriodicTaskMutationRequest } from './CeleryPeriodicTaskMutationRequest';
import {
    CeleryPeriodicTaskMutationRequestFromJSON,
    CeleryPeriodicTaskMutationRequestFromJSONTyped,
    CeleryPeriodicTaskMutationRequestToJSON,
} from './CeleryPeriodicTaskMutationRequest';

/**
 * 
 * @export
 * @interface PeriodicTaskMutationRequest
 */
export interface PeriodicTaskMutationRequest {
    /**
     * 
     * @type {number}
     * @memberof PeriodicTaskMutationRequest
     */
    repository?: number;
    /**
     * 
     * @type {CeleryPeriodicTaskMutationRequest}
     * @memberof PeriodicTaskMutationRequest
     */
    periodic?: CeleryPeriodicTaskMutationRequest;
    /**
     * 
     * @type {string}
     * @memberof PeriodicTaskMutationRequest
     */
    playbook?: string;
    /**
     * 
     * @type {string}
     * @memberof PeriodicTaskMutationRequest
     */
    inventories?: string;
    /**
     * 
     * @type {string}
     * @memberof PeriodicTaskMutationRequest
     */
    envvars?: string;
    /**
     * 
     * @type {string}
     * @memberof PeriodicTaskMutationRequest
     */
    extravars?: string;
    /**
     * 
     * @type {number}
     * @memberof PeriodicTaskMutationRequest
     */
    createdBy?: number | null;
    /**
     * 
     * @type {number}
     * @memberof PeriodicTaskMutationRequest
     */
    updatedBy?: number | null;
}

/**
 * Check if a given object implements the PeriodicTaskMutationRequest interface.
 */
export function instanceOfPeriodicTaskMutationRequest(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function PeriodicTaskMutationRequestFromJSON(json: any): PeriodicTaskMutationRequest {
    return PeriodicTaskMutationRequestFromJSONTyped(json, false);
}

export function PeriodicTaskMutationRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PeriodicTaskMutationRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'repository': !exists(json, 'repository') ? undefined : json['repository'],
        'periodic': !exists(json, 'periodic') ? undefined : CeleryPeriodicTaskMutationRequestFromJSON(json['periodic']),
        'playbook': !exists(json, 'playbook') ? undefined : json['playbook'],
        'inventories': !exists(json, 'inventories') ? undefined : json['inventories'],
        'envvars': !exists(json, 'envvars') ? undefined : json['envvars'],
        'extravars': !exists(json, 'extravars') ? undefined : json['extravars'],
        'createdBy': !exists(json, 'created_by') ? undefined : json['created_by'],
        'updatedBy': !exists(json, 'updated_by') ? undefined : json['updated_by'],
    };
}

export function PeriodicTaskMutationRequestToJSON(value?: PeriodicTaskMutationRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'repository': value.repository,
        'periodic': CeleryPeriodicTaskMutationRequestToJSON(value.periodic),
        'playbook': value.playbook,
        'inventories': value.inventories,
        'envvars': value.envvars,
        'extravars': value.extravars,
        'created_by': value.createdBy,
        'updated_by': value.updatedBy,
    };
}